sorry... lcm 0 0 =0

+++ float.cpp
@@ -178,7 +178,7 @@ Number *Float::lcm(Number *number2){
 	Number *tmp2 = number2->abs();
 	Float *tmp = SCAST_FLOAT(tmp2);
 	if (number_ == 0 && tmp->number_ == 0){
-		Float *result = new Float(1);
+		Float *result = new Float(0);
 		return result;
 	}
 	Float *result = SCAST_FLOAT(this->mul(number2)->div(gcd(number2)));
